remove debug code
